refactor: update path to import cookie name consts
