+++ src/app/intro/page.tsx
@@ -15,9 +15,11 @@ import {
   newMapIdStorage,
   nicknameStorage,
   invitationLinkStorage,
+  AUTHORIZATION,
 } from '@/utils/storage'
 import useCookie from '@/hooks/use-cookie'
 import { useIsServer } from '@/hooks/use-is-server'
+import { RECENT_MAP_ID } from '@/utils/storage/index'
 
 export interface IntroActionDispatch {
   goNextStep: VoidFunction
@@ -52,14 +54,14 @@ const Intro = () => {
   const isServer = useIsServer()
   const router = useRouter()
 
-  const recentMapId = useCookie('recent_map_id')
+  const recentMapId = useCookie(RECENT_MAP_ID)
   useEffect(() => {
     if (recentMapId) {
       router.push(`/map/${recentMapId}`)
     }
   }, [router, recentMapId])
 
-  const authorization = useCookie('Authorization')
+  const authorization = useCookie(AUTHORIZATION)
   const nickname = nicknameStorage.getValueOrNull()
   const newMapId = newMapIdStorage.getValueOrNull()
   const getInitialStep = useMemo(() => {
