+++ src/components/intro/steps/invite.tsx
@@ -6,7 +6,7 @@ import { Typography } from '@/components/common'
 import ConfirmCancelButton from '@/components/confirm-cancel-button'
 import { newMapIdStorage } from '@/utils/storage'
 import { setCookie } from '@/app/actions'
-import { RECENT_MAP_ID } from '@/utils/storage/cookie'
+import { RECENT_MAP_ID } from '@/utils/storage/index'
 import InvitingBoardingPass from '@/components/boarding-pass/inviting-boarding-pass'
 
 const Invite = () => {
